Refactor iOS credential matching to use shared method with UI logic (#1120)
