+++ apps/mobile-app/ios/VaultUI/Views/CredentialProviderView.swift
@@ -247,91 +247,7 @@ public class CredentialProviderViewModel: ObservableObject {
     }
 
    func filterCredentials() {
-        if searchText.isEmpty {
-            filteredCredentials = credentials
-            return
-        }
-
-        func extractDomain(from urlString: String) -> String {
-            var domain = urlString.lowercased().trimmingCharacters(in: .whitespacesAndNewlines)
-            // Remove protocol if present
-            domain = domain.replacingOccurrences(of: "https://", with: "")
-            domain = domain.replacingOccurrences(of: "http://", with: "")
-            // Remove www. prefix
-            domain = domain.replacingOccurrences(of: "www.", with: "")
-            // Remove path, query, and fragment
-            if let firstSlash = domain.firstIndex(of: "/") {
-                domain = String(domain[..<firstSlash])
-            }
-            if let firstQuestion = domain.firstIndex(of: "?") {
-                domain = String(domain[..<firstQuestion])
-            }
-            if let firstHash = domain.firstIndex(of: "#") {
-                domain = String(domain[..<firstHash])
-            }
-            return domain
-        }
-
-        func extractRootDomain(from domain: String) -> String {
-            let parts = domain.components(separatedBy: ".")
-            return parts.count >= 2 ? parts.suffix(2).joined(separator: ".") : domain
-        }
-
-        func domainsMatch(_ domain1: String, _ domain2: String) -> Bool {
-            let d1 = extractDomain(from: domain1)
-            let d2 = extractDomain(from: domain2)
-            
-            // Exact match
-            if d1 == d2 { return true }
-            
-            // Check if one domain contains the other (for subdomain matching)
-            if d1.contains(d2) || d2.contains(d1) { return true }
-            
-            // Check root domain match
-            let d1Root = extractRootDomain(from: d1)
-            let d2Root = extractRootDomain(from: d2)
-            
-            return d1Root == d2Root
-        }
-
-        // Try to parse as URL first
-        let searchDomain = extractDomain(from: searchText)
-        
-        if !searchDomain.isEmpty {
-            var matches: Set<Credential> = []
-
-            // Check for domain matches with priority
-            credentials.forEach { credential in
-                guard let serviceUrl = credential.service.url, !serviceUrl.isEmpty else { return }
-                
-                if domainsMatch(searchText, serviceUrl) {
-                    matches.insert(credential)
-                }
-            }
-
-            // If no domain matches found, try text search in service name
-            if matches.isEmpty {
-                let domainParts = searchDomain.components(separatedBy: ".")
-                let domainWithoutExtension = domainParts.first?.lowercased() ?? searchDomain.lowercased()
-                
-                let nameMatches = credentials.filter { credential in
-                    let serviceNameMatch = credential.service.name?.lowercased().contains(domainWithoutExtension) ?? false
-                    let notesMatch = credential.notes?.lowercased().contains(domainWithoutExtension) ?? false
-                    return serviceNameMatch || notesMatch
-                }
-                matches.formUnion(nameMatches)
-            }
-
-            filteredCredentials = Array(matches)
-        } else {
-            // Non-URL fallback: simple text search in service name, username, or notes
-            let lowercasedSearch = searchText.lowercased()
-            filteredCredentials = credentials.filter { credential in
-                (credential.service.name?.lowercased().contains(lowercasedSearch) ?? false) ||
-                (credential.username?.lowercased().contains(lowercasedSearch) ?? false) ||
-                (credential.notes?.lowercased().contains(lowercasedSearch) ?? false)
-            }
-        }
+        filteredCredentials = CredentialFilter.filterCredentials(credentials, searchText: searchText)
     }
